+++ src/main/java/tragicneko/tragicmc/worldgen/biome/BiomeGenCrystal.java
@@ -15,7 +15,7 @@ public class BiomeGenCrystal extends TragicBiome {
 	public CustomSpikesWorldGen crystalGen;
 
 	public BiomeGenCrystal(int par1) {
-		super(par1, 0);
+		super(par1, (byte) 0);
 		this.spawnableCreatureList.clear();
 		this.spawnableMonsterList.clear();
 		this.spawnableCaveCreatureList.clear();
@@ -31,7 +31,7 @@ public class BiomeGenCrystal extends TragicBiome {
 		this.theBiomeDecorator.treesPerChunk = 0;
 		this.fillerBlock = TragicBlocks.Crystal;
 		this.topBlock = TragicBlocks.Crystal;
-		this.crystalGen = new CustomSpikesWorldGen(8, TragicBlocks.Crystal, 0, 0.90477735D, 0.421114525D, 1.15D, 1.15D, false, false);
+		this.crystalGen = new CustomSpikesWorldGen((byte) 8, TragicBlocks.Crystal, (byte) 0, 0.90477735D, 0.421114525D, 1.15D, 1.15D, false, false);
 	}
 
 	@Override
